requirements.txt
dash==2.7.0
dash_bootstrap_components==1.2.1
pandas==1.5.2
pendulum==2.1.2
plotly==5.11.0
PyYAML==6.0
requests==2.28.1
termcolor==2.1.1
